requirements.txt
# Core dependencies
python-dotenv==1.0.0
aiohttp==3.9.1
websockets==12.0
requests==2.31.0

# Ethereum/Web3
eth-account==0.11.0
eth-utils==2.3.2
web3==6.11.3

# Telegram bot
python-telegram-bot==20.7

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Utilities
loguru==0.7.2
pydantic==2.5.0
msgpack==1.1.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
